lots of changes... Allow user to mount more than just sdcards (anything EXT) to PC's be sure to use the built in format for data in nandroid (so it doesnt kill /data/media if present)

+++ nandroid-mobile.sh
@@ -532,7 +532,11 @@ if [ "$RESTORE" == 1 ]; then
 	fi
 	echo "* print Erasing /$image..."
 	[ "$PROGRESS" == "1" ] && echo "* show_indeterminate_progress"
-	rm -rf /$image/*
+	if [ "$image" != "data" ]; then
+	  rm -rf /$image/*
+	else 
+	  format /$image
+	fi
 	
 	TAR_OPTS="x"
 	[ "$PROGRESS" == "1" ] && TAR_OPTS="${TAR_OPTS}v"
